added crime_level class, some additional parser code, many todo's
