--- src/backend/expungeservice/analyzer/__init__.py
@@ -1,1 +1,0 @@
-from .analyze import *
