+++ src/backend/expungeservice/models/charge.py
@@ -31,6 +31,16 @@ class Charge(object):
         self.date = date
         self.disposition = disposition
 
+        self._result = None
+
+        @property
+        def type_elig_result(self):
+            return self._result
+
+        @type_elig_result.setter
+        def type_elig_result(self, result):
+            self._result = result
+
         self.type_eligible = False # this is going to be a bool that represents if this is expungable
         self.time_eligible = False
 
@@ -39,3 +49,36 @@ class Charge(object):
 
     # def type_eligible(self):
     #     if self.statute
+
+
+
+# class Charge(object):
+#     """ Charge filed on a Client.
+#
+#     Attributes:
+#         name: A string describing the charge.
+#         statute: A Statute object that applies for the charge.
+#         date: A datetime.date object specifying the date of the charge.
+#         disposition: A Disposition object for the charge.
+#     """
+#     def __init__(
+#             self,
+#             name,
+#             statute,
+#             level,
+#             date,
+#             disposition):
+#         self.name = name
+#         self.statute = statute
+#         self.level = level
+#         self.date = date
+#         self.disposition = disposition
+#         self._result = None
+#
+#     @property
+#     def type_elig_result(self):
+#         return self._result
+#
+#     @type_elig_result.setter
+#     def type_elig_result(self, result):
+#         self._result = result
