+++ src/backend/expungeservice/models/disposition.py
@@ -1,20 +1,25 @@
 import enum
 
-class Disposition:
+class Disposition(object):
+    """ Disposition for a charge.
 
-    def __init__(self):
-        self.date = None
-        self.ruling = None
+    Attributes:
+        type_: An enum of type DispositionType.
+        date: A datetime.date specifying the date of the disposition.
+    """
+    def __init__(self, type_, date):
+        self.type_ = type_
+        self.date = date
 
 
-# DispositionType = enum.Enum('Disposition',
-#                             ' '.join([
-#                                 'CONVICTION',
-#                                 'NO_CONVICTION',
-#                                 'DISMISSED',
-#                                 'ACQUITTED',
-#                                 'NO_COMPLAINT'
-#                                 ]))
+DispositionType = enum.Enum('Disposition',
+                            ' '.join([
+                                'CONVICTION',
+                                'NO_CONVICTION',
+                                'DISMISSED',
+                                'ACQUITTED',
+                                'NO_COMPLAINT'
+                                ]))
 #
 # class Disposition(object):
 #     """ Disposition for a charge.
@@ -26,3 +31,5 @@ class Disposition:
 #     def __init__(self, type_, date):
 #         self.type_ = type_
 #         self.date = date
+
+
