+++ src/backend/expungeservice/models/statute.py
@@ -5,6 +5,10 @@
 #todo: find out if section and subsection are relevant
 from expungeservice.analyzer.ineligible_crimes_list import IneligibleCrimesList
 
+
+
+
+
 class Statute(object):
     """ Statute corresponding to a law
 
@@ -25,6 +29,7 @@ class Statute(object):
         self.subchapter = subchapter
         self.section = section
         self.subsection = subsection
+        # TODO we may need to add components beyond subsection
 
         if len(statute_string)>=6: #todo: this is wrong but will kinda work for everything on our list except marijuana crimes.
 
@@ -40,8 +45,27 @@ class Statute(object):
 
             self.statute_string = self.chapter + '.' + self.subchapter
 
+    def __eq__(self, other):
+        return (self.chapter == other.chapter and
+                self.subchapter == other.subchapter and
+                ((not self.section and not other.section) or
+                 self.section == other.section) and
+                ((not self.subsection and not other.subsection) or
+                 self.subsection == other.subsection))
+
     def __str__(self):
         return str(self.statute_string)
+
+    # def __str__(self):
+    #     # TODO do these need to have leading zeros?
+    #     statute = '{}'.format(self.chapter)
+    #     if self.subchapter:
+    #         statute = '{}.{:03d}'.format(statute, self.subchapter)
+    #     if self.section:
+    #         statute = '{}({})'.format(statute, self.section)
+    #     if self.subsection:
+    #         statute = '{}({})'.format(statute, self.subsection)
+    #     return statute
